+++ tests/testthat/test_applyLookup_expandVar.R
@@ -16,10 +16,6 @@ test_that("Errors for apply lookup with expanding into multiple variables", {
   expect_error(applyLookup_expandVar(l_gads, lookup2),
                "'lookup' table has to be formatted correctly.")
 
-  lookup3$value[1:2] <- NA
-  expect_error(applyLookup_expandVar(l_gads, lookup3),
-               "In more than 1 row value is missing.")
-
   lookup4$new_value1[1] <- NA
   expect_warning(applyLookup_expandVar(l_gads, lookup4),
                  "Not all values have a recode value assigned (missings in value_new).", fixed = TRUE)