+++ R/checkFormat.r
@@ -28,14 +28,21 @@ checkFormat <- function(GADSdat, type = "SPSS", changeFormat = TRUE) {
 checkFormat.GADSdat <- function(GADSdat, type = "SPSS", changeFormat = TRUE) {
   check_GADSdat(GADSdat)
   labels <- GADSdat$labels
+  anyUnknownFormat <- FALSE
 
   naNam <- names(GADSdat$dat)[which(unlist(lapply(GADSdat$dat, function(x) all(is.na(x)))))]
   for(hh in naNam) {
-    if(grepl("^A", labels$format[labels$varName==hh][1]) & (is.numeric(GADSdat$dat[,hh]) | is.logical(GADSdat$dat[,hh]))) {
-      GADSdat$dat[,hh] <- as.character(GADSdat$dat[,hh])
+    if(!is.na(labels$format[labels$varName==hh][1])) {
+      if(grepl("^A", labels$format[labels$varName==hh][1]) & (is.numeric(GADSdat$dat[,hh]) | is.logical(GADSdat$dat[,hh]))) {
+        GADSdat$dat[,hh] <- as.character(GADSdat$dat[,hh])
+      } else {
+        if(grepl("^F", labels$format[labels$varName==hh][1]) & (is.character(GADSdat$dat[,hh]) | is.logical(GADSdat$dat[,hh])))
+        GADSdat$dat[,hh] <- as.numeric(GADSdat$dat[,hh])
+      }
     } else {
-      if(grepl("^F", labels$format[labels$varName==hh][1]) & (is.character(GADSdat$dat[,hh]) | is.logical(GADSdat$dat[,hh])))
-      GADSdat$dat[,hh] <- as.numeric(GADSdat$dat[,hh])
+     message("Format and data for variable '", hh, "' are NA. Format cannot be derived from data and will be set to A1.")
+      GADSdat$dat[,hh] <- as.character(GADSdat$dat[,hh])
+      anyUnknownFormat <- TRUE
     }
     }
 
@@ -54,6 +61,8 @@ checkFormat.GADSdat <- function(GADSdat, type = "SPSS", changeFormat = TRUE) {
     }
   })
 
+  if(isTRUE(anyUnknownFormat)) lengths[lengths == -Inf] <- 1
+
   decimals <- sapply(names(GADSdat$dat), function(ll) { if(isTRUE(is.numeric(GADSdat$dat[,ll]) &
                                                                   all(is.numeric(utils::type.convert(labels$value[labels$varName==ll],as.is=TRUE))|
                                                                                                      is.na(labels$value[labels$varName==ll])))) {
@@ -66,6 +75,8 @@ checkFormat.GADSdat <- function(GADSdat, type = "SPSS", changeFormat = TRUE) {
     }
   }
   })
+
+  if(isTRUE(anyUnknownFormat)) decimals[decimals == -Inf] <- 1
 
   varsWithDecimals <-  names(which(lengths != decimals))