+++ tests/testthat/test_write_spss2.R
@@ -1,5 +1,5 @@
-g <- import_raw(df = data.frame(var1 = c("ab c","bb"), var2 = c(1,NaN), var3 = c(1.01, 2.00), var4 = c(1.0001, 4.00187243564195786431587643596),var5 = c("ab c","b\"b"),var6 = c("ab c","b\"b"),stringsAsFactors = FALSE),
-                varLabels = data.frame(varName = c("var1", "var2", "var3", "var4","var5","var6"), varLabel = c("a label", NA, "another label", NA, NA, NA), stringsAsFactors = FALSE),
+g <- import_raw(df = data.frame(var1 = c("ab c","bb"), var2 = c(1,NaN), var3 = c(1.01, 2.00), var4 = c(1.0001, 4.00187243564195786431587643596),var5 = c("ab c","b\"b"),var6 = c("ab c","b\"b"),var7=NA,stringsAsFactors = FALSE),
+                varLabels = data.frame(varName = c("var1", "var2", "var3", "var4","var5","var6","var7"), varLabel = c("a label", NA, "another label", NA, NA, NA,NA), stringsAsFactors = FALSE),
                 valLabels = data.frame(varName = c("var1", "var1", "var2", "var2", "var2","var3", "var3","var5","var6"), value = c(-96, -99, -99, 0, 1,-96, -99, -9,-9),
                                        valLabel = c("miss1", "miss2", "miss2","right","wrong","miss1", "miss2", NA, NA),
                                        missings = c("miss", "miss", "miss","valid","valid","miss", "miss", "miss", "miss"), stringsAsFactors = FALSE))
@@ -7,7 +7,11 @@ g <- import_raw(df = data.frame(var1 = c("ab c","bb"), var2 = c(1,NaN), var3 = c
 
 f_txt <- tempfile(fileext = ".txt")
 f_sps <- gsub("txt$", "sps", f_txt)
-write_spss2(g, txtPath = f_txt, dec=",")
+
+test_that("writeData", {
+  expect_message(write_spss2(g, txtPath = f_txt, dec=","), "Format and data for variable 'var7' are NA. Format cannot be derived from data and will be set to A1.", fixed=TRUE)
+})
+
 syntax <- readChar(f_sps, file.info(f_sps)$size)
 
 
@@ -45,8 +49,8 @@ test_that("Format and writeHeader", {
 
 test_that("writeData", {
   expect_message(g1<-writeData(g, txtPath = f_txt, dec=",", fileEncoding = "UTF-8"),"In character variable(s) var5, var6 quotation marks (\") had to be replaced with inverted commas (').", fixed=TRUE)
-  expect_equal(g1$dat, data.frame(var1 = c("ab c","bb"), var2 = c(1,NaN), var3 = c(1.01, 2.00), var4 = c(1.0001, 4.00187243564195786431587643596),var5 = c("ab c","b'b"),var6 = c("ab c","b'b"),stringsAsFactors = FALSE))
-  expect_equal(read.csv2(f_txt, header=FALSE), data.frame(V1 = c("ab c","bb"), V2 = c(1,NaN), V3 = c(1.01, 2.00), V4 = c(1.0001, 4.00187243564195786431587643596),V5 = c("ab c","b'b"),V6 = c("ab c","b'b"),stringsAsFactors = FALSE))
+  expect_equal(g1$dat, data.frame(var1 = c("ab c","bb"), var2 = c(1,NaN), var3 = c(1.01, 2.00), var4 = c(1.0001, 4.00187243564195786431587643596),var5 = c("ab c","b'b"),var6 = c("ab c","b'b"),var7=NA,stringsAsFactors = FALSE))
+  expect_equal(read.csv2(f_txt, header=FALSE), data.frame(V1 = c("ab c","bb"), V2 = c(1,NaN), V3 = c(1.01, 2.00), V4 = c(1.0001, 4.00187243564195786431587643596),V5 = c("ab c","b'b"),V6 = c("ab c","b'b"),V7=NA,stringsAsFactors = FALSE))
 })