+++ NEWS.md
@@ -5,6 +5,7 @@
 ## bug fixes
 * `applyChangeMeta()` and `recodeGADS()` now correctly perform recodings (and throw errors) if multiple meta data conflicts occur (#57)
 * `removeEmptyValLabels()` for removing unused missing tags and value labels (#4)
+* `extractData()` and `extractData2()` now correctly apply value labels, even when value and value label conflicts exist
 
 # eatGADS 1.1.0