+++ NEWS.md
@@ -3,6 +3,7 @@
 * `inspectDifferences()` and `inspectMetaDifferences()` now allow comparisons of variables within the same `GADSdat` object (#62)
 
 ## bug fixes
+* `applyLookup()` now works for multiple variables that contain `NAs` (#68)
 * `applyChangeMeta()` and `recodeGADS()` now correctly perform recodings (and throw errors) if multiple meta data conflicts occur (#57)
 * `removeEmptyValLabels()` for removing unused missing tags and value labels (#4)
 * `extractData()` and `extractData2()` now correctly apply value labels, even when value and value label conflicts exist