+++ NEWS.md
@@ -10,7 +10,13 @@
 * `write_spss2()` now also handles variables containing only NA values when `format` is also NA (#72)
 * `changeMissings()` now correctly changes missing labels for values of variables with partially non-existent `value`s and/or `valLabel`s (#73)
 * `autoRecode()` now correctly overwrites the existing variable if `var_suffix = ""` (#84)
-* `extractData2()` now correctly transforms variables with duplicate value labels (#77)
+* `extractData2()` and `extractData()` now correctly transform variables with duplicate value labels (#77)
+
+## documentation
+* `import_spss` vignette updated to use `extractData2()` instead of `extractData()`
+
+## internal
+* refactored `extractData()` to use `extractData2()` internally
 
 # eatGADS 1.1.0