+++ NEWS.md
@@ -10,6 +10,7 @@
 * `write_spss2()` now also handles variables containing only NA values when `format` is also NA (#72)
 * `changeMissings()` now correctly changes missing labels for values of variables with partially non-existent `value`s and/or `valLabel`s (#73)
 * `autoRecode()` now correctly overwrites the existing variable if `var_suffix = ""` (#84)
+* The output of `inspectMetaDifferences()` is now correctly named even if differences in variable labels and `SPSS` format occur (#81)
 * `extractData2()` and `extractData()` now correctly transform variables with duplicate value labels (#77)
 
 ## documentation